pyproject.toml
[tool.poetry]
name = "dazbog"
version = "0.1.0"
description = "Generate types and resolvers for ariadne from grapql schema"
authors = ["Damian Świstowski <damian@swistowski.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<4"

click = "^8.0.3"
graphql-core = "^3.1.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.10b0"
ariadne = "^0.13.0"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
